% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LCPrincple}
\alias{LCPrincple}
\title{Least-Committed  Principle for creating bbas}
\usage{
LCPrincple(Mat)
}
\arguments{
\item{Mat}{matrix, \eqn{m \times k}, \eqn{m} is the number of sources, \eqn{k} is the length of probability vectors. If the number of sources is 1, the input probability could be a vector.}
}
\value{
mass_bba matrix, \eqn{m \times 2^k}, each column is a bba. If there is only one source, the output is a bba vector.
}
\description{
Least-Committed  Principle for creating bbas
}
\examples{
pro1 = c(0.25, 0.25, 0.25, 0.25);
pro2 = c(0.3, 0.2, 0.2, 0.1);
pro3 = rbind(pro1, pro2);

LCPrincple(pro1)
LCPrincple(pro2)
LCPrincple(pro3)
}

