% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LCRule}
\alias{LCRule}
\title{Least-Committed  Rules for creating bbas}
\usage{
LCRule(Mat)
}
\arguments{
\item{Mat}{matirx, \eqn{m \times k}, \eqn{m} is the number of sources, \eqn{k} is the length of probability vectors. If the numer of sources is 1, the input probabiltiy could be a vector.}
}
\value{
mass_bba matrix, \eqn{m \times 2^k}, each column is a bba. If there is only one source, the output is a bba vector.
}
\description{
Least-Committed  Rules for creating bbas
}
\examples{
pro1 = c(0.25, 0.25, 0.25, 0.25);
pro2 = c(0.3, 0.2, 0.2, 0.1);
pro3 = rbind(pro1, pro2);

LCRule(pro1)
LCRule(pro2)
LCRule(pro3)
}

